BJS2-79034. Обновление структуры проекта после мерджа GoalService
1) Убраны дублирующие классы
2) Исключения GoalService добавлены в GlobalExceptionHandler
3) Добавлено уникальное исключения для отсутсвия id пользователя в заголовке запроса
4) Добавлены недостающие директории
5) Унифицировано расположение типовых классов
