+++ src/main/java/school/faang/user_service/controller/event/EventControllerImpl.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.controller;
+package school.faang.user_service.controller.event;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
