+++ src/main/java/school/faang/user_service/dto/goal/request/UpdateGoalDto.java
@@ -1,18 +1,19 @@
-package school.faang.user_service.entity.goal.dto.request;
+package school.faang.user_service.dto.goal.request;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import jakarta.validation.constraints.Future;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
+import school.faang.user_service.entity.goal.GoalStatus;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public record CreateGoalDto(
+public record UpdateGoalDto(
         @NotBlank(message = "Empty goal title not allowed!") String title,
         String description,
-        Long parentId,
+        GoalStatus status,
         @NotNull List<Long> skillsId,
-        @Future(message = "Deadline must be in future!") LocalDateTime deadline
+        @Future(message = "Dead line must be in future!") LocalDateTime deadline
 ) {}
