+++ src/main/java/school/faang/user_service/exception/common/DataValidationException.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.exception;
+package school.faang.user_service.exception.common;
 
 public class DataValidationException extends RuntimeException {
     public DataValidationException(String message) {
