+++ src/main/java/school/faang/user_service/exception/common/PreConditionFailedException.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.exception;
+package school.faang.user_service.exception.common;
 
 public class PreConditionFailedException extends RuntimeException {
     public PreConditionFailedException(String message) {
