+++ src/main/java/school/faang/user_service/exception/common/RecordNotFoundException.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.exception;
+package school.faang.user_service.exception.common;
 
 public class RecordNotFoundException extends RuntimeException {
     public RecordNotFoundException(String message) {
