+++ src/main/java/school/faang/user_service/exception/event/EventCreationNotAllowedException.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.exception;
+package school.faang.user_service.exception.event;
 
 public class EventCreationNotAllowedException extends RuntimeException {
     public EventCreationNotAllowedException(String message) {
