+++ src/main/java/school/faang/user_service/exception/users/UserIdNotFoundException.java
@@ -1,0 +1,7 @@
+package school.faang.user_service.exception.users;
+
+public class UserIdNotFoundException extends RuntimeException {
+    public UserIdNotFoundException(String message) {
+        super(message);
+    }
+}
